Added basic Fuel and Hull (health) UI and functionality
Replaced the async DigAnimation by a coroutine. async was throwing an error when we closed the game, but a coroutine doesn't (and async is hard to stop, a coroutine is easy to stop)

Reduced Jitter by half when digging

Added Idle Fuel Consumption
Added Fuel Consumption when digging
Added Fuel Consumption when moving

Added Fall Damage, but we take damage from any sides, not just from falling. It's more like "Hit Damage" at the moment.