Trying to make some stuff less complicated
.. or at least I tried

Added a public static reference to AllTilesInfo rather than having to go through the GameManager every time

Removed the public static reference to the InventoryUI since it is only used in one other script, and passed the reference through a SerializeField instead.

Changed from automatically adding Ore Inventory Slots to the Ore Inventory UI to adding them manually. This allows for seeing the actual result in the engine

Added a debugging disable option on the player for fuel and health

Renamed a few variables and methods