Max speed for digging, fixed "floating offset"
Added maximum movement speed to dig.

Fixed "floating offset" when digging left/right due to moving the player to the center of the tile.
We now only move to the tile's X position when going left/right. When digging down, we also subtract 1 to Y pos.

For some reasons, I had to reduce the Player BoxCollider2D's Y size to 0.96666667 to visually match the floor contact. Weird.