Updated inventory UI, disableDigAnimation "upgrade"
Added ore value and weight to inventory UI
Value is individual, might want to change to total instead.

It was a pain to update, since changing the UI has to be done in the prefab, within which we don't have the proper alignment.

Added small debug feature where if we disableDigAnimation, we ignore the maximum digging speed.
Allows to continuously dig as fast as we can move.
This was "broken" since the addition of max digging speed, so this fix merely restores functionality.