AllTilesInfo isn't static anymore, fixed Inventory UI ref issue
AllTilesInfo isn't static anymore, we're instead passing the AllTilesInfo ScriptableObject as a reference where needed
Not sure if that's better or worse to be honest.

Moved OreInventorySlot.cs' gameManager and oreInventory initialization to Awake(), since there was a problem the first time the inventory was open where oreInventory was a None reference.
Probably an order of operation issue.