Added Pause Menu, with Resume/Quit
There is a annoyance where if we have the Upgrades Store UI open, click on a button (which sets the focus), then Pause the game, the focus is still in the Upgrades UI. The arrows and space buttons still controls the UI in the background.
I'll have to change the focus - but I want to change the Input system to the new, well, Input System.
That'll probably change the way to fix that issue.