Replaced legacy Input with the new Input System
Replaced legacy Input system with the new Input System.
Moved input management to FixedUpdate()

We cannot close the Upgrades UI by pressing "e", which is annoying since that's the only key that can open it. We need to press "Escape".
The issue is within the Menu Controls mode, if we include "e" in Cancel (or with a new Interact Action), then when the game is Paused, we could still interact in the background.
We technically need a new Action Map that distinguish Menu Controls to the Upgrades UI controls.
We'd have twice the Menu controls, is that really the fix ??

We can't Pause the game while in a "locking" Menu, by design - might want to change that later. But for now, the player is expected to press Escape once to close the UI, then again to Pause the game.
The Upgrades Store is an example of this, while the Inventory isn't affected.

I don't understand the point of the EventSystem's Input System UI Input Module. If I don't have it, I can still open/close the menu with Escape, but I can't interact with any menu using the mouse.
Does that mean I'd need to implement "Point" through script, if I were to not use the aforementioned Module ?
The Input Debug Window shows double the MenuControls Actions when the Module is fully populated and we're in Menu mode. In Gameplay mode, I have both Gameplay and Menu Actions.