Incorporated Compute Shader project
Imported another project where I played with shaders and Compute Shaders with the goal to develop the technology to be used for the miner game within the Level Generation and Tiles Rendering systems.

The idea is to use a Compute Shader (CS) to generate a texture where each pixel is a tile, with the pixel's color representing the tile type.
Then, we would sample the texture by doing a World Space to Texture Space conversion and convert the color to a TileID.
We would also have a second texture to represent the Tiles that were dug up.

We'd have a script to do the World->Texture space conversion, texture lookup, color->TileID conversion, etc.

The TileID texture lookup wouldn't be super necessary on its own, we can easily compute TileID through C# on the fly.
The actual reason we need a TileIDs texture is for it to be passed into the Rendering pipeline.
Instead of having one quad for each tiles, we would have a single quad, the shader of which handles all the visual required.
Picking the proper texture to display for each TileIDs (should be an Atlas with all the tile types, matching TileIDs' ordering), hiding dug up tiles, rounding up corners around dug up tiles, perhaps even animating a tile eroding away as the player digs it up.

The TileIDs texture only need to contain what is visible on-screen. (it would probably be larger, for debug sake ?)
At depth 500, we don't need to know what's above 450, for instance. (unless we can see more than 50 lines, which we should not)
The Tiles Dug Up texture will require to contain all the tiles, all the time. Starts empty, then each new depth layer is more Y resolution. (or perhaps we only increase the size by X each X depth)

We still need to compute the collision. We only need the Tiles Dug Up texture for that.

We will probably be able to remove the ray intersection to fetch the tile, since this would be replaced with the aforementioned script to handle the conversions and lookups.