Restructured "Sample from ComputeShader.cs"
Made the script more friendly to be used by other scripts.
Removed the code from Update() and into specific methods that can be called from elsewhere.

Added an example script (SampleExample.cs) to show how to fetch color properly.

If multiple SampleExample scripts are in the level, only a single will be used to update the Visualizer - but both will properly sample the color.
As far as functionality goes, this is no concern.
.. Although, it would be nice to be able to "properly" visualize the sample location at each of the locations we would want, instead of a single one.