Replaced idx/idy with single pixelID
Replaced idx/idy with a single pixelID.

Implemented caching in a way to be handled by each script sampling from the Compute Shader.
Perhaps it's not the greatest to expect each scripts to implement caching themselves, but I don't know how to make a generalized solution, so that's what we're going with at the moment.
An advantage to caching is the last thing to sample will be the visualized location within the Compute Shader.