Explosive and first pixel mapping exploration
The code is in a broken state.
It was written months ago, I don't recall much.
Looks like this was the first exploration of pixel mapping for the world generation idea where we'd make a texture in which each pixel is a block, offloaded to the GPU with a Compute Shader.
GameManager.PositionToPixelID() seems to be an attempt at this - but again, the code isn't working.
I'm submitting this as a backup to start from when refactoring or something.

There seems to also be a first implementation of a shortcut for explosives, but not much more than that.