Reversed LineGeneration's air_tiles[] logic
air_tiles[] was previously True for a tile and false for air, which I find super confusing.
True is now air, False is now a tile.

Changing was a little complex due to GenerateCollision2D().
Turns out the complexity was the default value of air_tiles[]. The last, additional element had a special meaning that when changing the true/false logic around, I had not considered. Line 28 fixes this.
Considering that, GenerateCollision2D() wasn't that hard to do, it's just inverting a few true to false and a few false to true.