Alternative collision handling, deeper pixel_ID integration
Added TilesCollisionAroundPlayer.cs containing 8 collision squares that snaps to place around the player.
This setup currently coexists with the previous system.
The previous one generates collision across all the visible level, the new system generates it around the player only.

Instead of passing a tile in scripts (say, to dig it), we use pixel_ID.
Previously, we had to pass the specific tile hit by the dig raycast to do stuff on it. Now, we pass the tile_ID, removing the dependency on individual tiles.

Moved logic to retrieve a specific TileInfo into GameManager.cs
Moved DigTile() logic into GameManager.cs

A bit of cleanup in PlayerScript.cs  -> Item_Explosive()
Since we're now using pixel_ID, we don't need to do a raycast to get the tiles around the player. We can get pixel_ID with an offset.