disableDigAnim tweaks, floor under trigger, tile xID removal
Removed now-unused xID property from TileScript.cs
Yet another step to remove functionality from TileScript.cs towards the potential future to ditch individual tiles in favor of a single quad for the entire level.

Added a check in CanPlayerDig() to return true if the dig animation is disabled. This allows to dig much faster when debugging.
Also, modulated the raycast length to know if there's a tile to gid near based on disableDigAnimation. When disabled, it's 0.6 (as previously set), but when enabled, it's 1.0. This is only to overcome framerate.

Added a trigger under the floor to disable it if the player ever gets under the floor while it's still active, somehow.
I don't know when that could happen, but during development, I sometimes move the player under the floor manually, and I got stranded not being able to go over ground. Easy enough to manually move the player back up, but this fix is also easy, so why not.