MoveAwayFromWall() and IsPlayerOnGround() tweaks
Fixed an issue introduced in previous submit where we were able to dig the walls without being grounded when the player was within the new floor trigger.
We now ignore triggers when fetching the floor.

New MoveAwayFromWall()
When falling/flying, the air->tile transition (i.e. where there's air gaps between tiles) sometimes (<1%) falsely detects a collision when the player is sliding right along a wall.
To fix this, we now force a small gap between the player and the wall.

Added lateral drag so the player won't move from one side of the screen to the other while flying up with a single tap of lateral movement.

Updated IsPlayerOnGround() logic.
We now sample the ground twice, on both sides under the player.
It's still not quite perfect, we sometimes get false positive when hitting a wall fast - but I don't think that will be an issue for the use cases of this function.