Enhanced player collision smoothness
Added a Look Up Table (LUT) for all 256 possibilities of collision shapes around the player.
Based on which tile isn't air, we fetch and apply the right collision arrangement.
The LUT is written in-code by hand, maybe we should store it in a separate file ?
I wonder if the current method is more efficient that if we did it programmatically ?
Each arrangements were made in a specific way to reduce "stray collisions" (aka "slip bumps").
The following arrangement
111
0P0 (P for Player)
000
Instead of three squares, we use a single one
[111]
 0P0
 000
This reduces the amount of squares required to represent the collision, as well as reducing the concern of hitting those "stray collisions".

The current LUT numbers do not consider the 0.05 transition padding on the X axis. They will be explored in the next submit.

There are two hacks that I'd like to fix, related to when the player above ground.
I believe the fix need to be done in GameManager.cs, within the various PixelIDTo__() methods.