TilesCollisionAroundPlayer overhaul
Yet another major update.
We enforce the previously-introduced concept of lut_number, which to reiterate, is a "collapsed" binary representation of un-dug tiles around the player
0 is no tile, 255 is all tiles, 165 is the four corners, 189 is two vertical lines of tiles, 231 is two horizontal lines, etc

A bunch of helper methods were added to handle lut_number.
GetBinaryList() takes the integer and turns it into a list.
39 -> [1,1,1,0,0,1]
2 -> [0,1]
248 -> [0,0,0,1,1,1,1,1]

ComputeLUTEntry() is the algorithm I couldn't be bothered to figure out in the previous submit.
It had taken me a few hours to write all 256 collision_LUT[][] elements in the previous submit. After that, I remembered I needed to add a gap for most of them, and just like that, a few more hours of tedious work appeared.
Not wanting to do that, I then coded this method in 1-2H, and it was fun to figure out.
Logic-ing was better than brute-forcing the solution. Figures.

I removed the LUT data for now though, since it's simpler to compute it on the fly. And it was driving VS nuts as it was trying to syntax highlight and give code suggestions.
If the "full algorithm" becomes a runtime performance concern, I can always regenerate the LUT using PrintLUT().
The implementation code is left in comments.

GetLUTNumberAroundPixelID() is the core of it all. It computes the lut_number based on which tiles are air or not around a certain pixel_ID (which happens to be the player for us).

Current issue is above ground, we're getting garbage data from GetLUTNumberAroundPixelID() due to GameManager.GetPixelIDAtOffset() not handling negative pixel_ID.
We currently force a certain LUT to mitigate the issue, but it makes the first layer (when above ground) always full.
Must fix, at some point.