Items implementation
Added implementation for six items
- Explosive -
Removes tiles around the player in a radius of 1.
- Large Explosive
Removes tiles around the player in a radius of 2.
- Cheap Teleporter
Teleports the player over ground with a random velocity.
- Fancy Teleporter
Safely teleports the player over ground, next to the fuel station, with no velocity.
- Repair Kit
Repairs the player's hull a fixed amount.
- Fuel Reserve
Refuels the player's tank a fixed amount.

Each items' implementation is contained in Items.cs
Each item is a class of its own, deriving from Item{}
They are instanced in Items{} in Start().
We've done away with ScriptableObjects. I never understood them back then, and I still don't.
Instead, Items is placed onto the PlayerObject and referenced into it. GameManager.cs does not interact with any item at the moment.

Items currently cannot be bought at the store, implementation is not done for that.

Removed the custom explosive logic in PlayerScript.cs, that sort of logic needs to live in the appropriate Item class.

Added shortcuts for each items, 1-6.
Shortcut key will certainly change later.
Bound each shortcut to the items within PlayerScript.cs

The steps to add a new item is at the top of Items.cs
The procedure is subject to change, which is why it isn't defined here. Also for discoverability.

Capped SetFuel() and SetHealth() to their respective maximums.