Item Store, thinking out loud about implementation
Laying out thoughts in Store_Items.cs about what sort of implementation could be made to buy items.

Also added an array containing each items. The idea is that this array would be what's referenced in other scripts when they need to handle all items, such as the store, save game, or other.
The array indices could be the items' order in the store.