Rock tile implementation
Added a rock tile scriptable object
I'm really not convinced that this is a great way to do this, but the alternative is to define all properties in a .cs file. Not sure if that would be good or bad.

Added a check to validate if we have a high-enough drill level to drill through tiles (only applicable to rocks at the moment).

Removed the individual hard-coded TileInfo entries on allTilesInfo, it's now an array in the inspector.
This allowed to remove the GetAllTIles() method. It defined the priority order of which tiles gets chosen by the algorithm; this is now specified by the inspector array.