Added copper ore
The full "how to add a new tile type" is shown in this submit.

Word of caution - adding an entry to the TileType enum creates an issue for existing tiles.
The scriptable objects save the TileType as an ID. dirt is 0, coal is 1, etc. If I add a new tile between dirt and coal, dirt will still be 0, but the new tile will be 1, coal will be 2, etc.